chore(NotificationSender): Move composable to separate file
